// src/interfaces/IconTypes.ts
export type ICON_TYPES =
  | "after effects"
  | "angular"
  | "blender"
  | "cpp"
  | "docker"
  | "electron"
  | "figma"
  | "firebase"
  | "gamemaker"
  | "graphql"
  | "illustrator"
  | "java"
  | "javascript"
  | "keras"
  | "mongodb"
  | "mysql"
  | "nodejs"
  | "opengl"
  | "photoshop"
  | "premiere"
  | "python"
  | "react"
  | "redis"
  | "substance_designer"
  | "substance_painter"
  | "tensorflow"
  | "threejs"
  | "typescript"
  | "unity"
  | "unreal"
  | "vulkan"
  | "webgl"
  | "xd";